% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_block_demographics.R
\name{generate_block_dmgs}
\alias{generate_block_dmgs}
\title{Generate demographic estimates given the block-level address counts}
\usage{
generate_block_dmgs(
  acs_data,
  bac_data,
  block_measure = "total_housing_units",
  acs_var_list = .GlobalEnv$named_acs_var_list
)
}
\arguments{
\item{acs_data}{ACS data from get_data.}

\item{bac_data}{Block-level address counts from get_data.}

\item{block_measure}{Measure to use from the bac_data}

\item{acs_var_list}{Named list of ACS Codes.}
}
\description{
Generate demographic estimates given the block-level address counts
}
