% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get ACS demographics and Census Block Address Counts for a County}
\usage{
get_data(state_abbrev = "", county_fips = "", geo_level = "block group")
}
\arguments{
\item{county_fips}{3 number FIPS Code.}

\item{geo_level}{The geographic level at which to estimate populations.}

\item{state_abbrv}{2-letter state abbreviation.}
}
\description{
Get ACS demographics and Census Block Address Counts for a County
}
\examples{
mydata <- get_data(state_abbrev = "VA", county_fips = "013")
}
