% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_acs_variables.R
\name{set_acs_variables}
\alias{set_acs_variables}
\title{Sets the ACS variables that will be used by get_data and generate_block_demographics.}
\usage{
set_acs_variables(named_acs_var_list = "")
}
\arguments{
\item{named_acs_var_list}{Optional. Named list of ACS variables.}
}
\description{
Sets the ACS variables that will be used by get_data and generate_block_demographics.
}
