% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_block_demographics.R
\name{generate_block_dmgs}
\alias{generate_block_dmgs}
\title{Generate demographic estimates given the block-level address counts}
\usage{
generate_block_dmgs(
  acs_data,
  bac_data,
  block_measure = "total_housing_units",
  acs_var_list = list(total_pop = "B01001_001", wht_alone = "B02001_002",
    afr_amer_alone = "B02001_003", male = "B01001_002", male0_4 = "B01001_003", male5_9 =
    "B01001_004", male10_14 = "B01001_005", male15_17 = "B01001_006", female =
    "B01001_026", female0_4 = "B01001_027", female5_9 = "B01001_028", female10_14 =
    "B01001_029", female15_17 = "B01001_030")
)
}
\arguments{
\item{acs_data}{ACS data from get_data.}

\item{bac_data}{Block-level address counts from get_data.}

\item{block_measure}{Measure to use from the bac_data}

\item{acs_var_list}{Named list of ACS Codes.}
}
\description{
Generate demographic estimates given the block-level address counts
}
\examples{
library(dc.census.block.address.counts)
mydata <- get_data(state_abbrev = "VA", county_fips = "013")
bk_dmgs <- generate_block_dmgs(acs_data = mydata$acs_data_51013, bac_data = mydata$bac_data_51013)
}
